Extend device permissions
A user reported [over on the TFE ticket for the Flatpak distribution](https://github.com/luciusDXL/TheForceEngine/issues/268#issuecomment-1960163198) that game controllers won't work with the TFE Flatpak without overriding the devices permission. As this is really a feature that should work out of the box, we should extend this so users can use gamepads without needing to manually override it.